rename presenter to controller
moved controller router into controller folder
seperate router, handlers and helpers in controller folder
