+++ controller/handler/trial.js
@@ -9,7 +9,8 @@ const moment = require('moment');
 const _ = require('lodash');
 const color = require('colors.css');
 
-const database = require('../model');
+const database = require('../../model');
+const processPatient = require('../helper/process-patient');
 
 /**
  * A dashboard with an overview of a specific trial.
@@ -69,7 +70,6 @@ function processTrial (currentTrial) {
     const endDate = moment(trial.endAt);
 
     return {
-
         id: trial.id,
         name: trial.name,
         description: trial.description,
@@ -84,36 +84,5 @@ function processTrial (currentTrial) {
         activePatients: Math.floor(Math.random() * 50),
         completedPatients: Math.floor(Math.random() * 40),
         compliantCount: Math.floor(Math.random() * 80)
-    };
-}
-
-/**
- * Takes in a Patient model and processes it into human readable format
- * @param {Trial} currentPatient - a single Patient object
- * @returns {Object} processed Patient
- */
-function processPatient (currentPatient) {
-    const patient = currentPatient.dataValues;
-    const statuses = ['Compliant', 'Semicompliant', 'Noncompliant'];
-    const statusTypes = ['success', 'warning', 'danger'];
-
-    // TODO replace randomly generated data with real data
-    const randomStatus = Math.floor(Math.random() * 3);
-    const randomStage = Math.floor(Math.random() * 3);
-    const randomMissed = Math.floor(Math.random() * 10);
-    const randomConsecutiveMissed = Math.floor(randomMissed / 3);
-    const startDate = new Date(1, 1, 2014);
-    const todayDate = new Date();
-    const randomDate = new Date(startDate.getTime() + Math.random() * (todayDate.getTime() - startDate.getTime()));
-    const randomDateDisplay = moment(randomDate).format('L');
-
-    return {
-        pin: patient.pin,
-        status: statuses[randomStatus],
-        statusType: statusTypes[randomStatus],
-        stage: randomStage,
-        lastTaken: randomDateDisplay,
-        totalMissed: randomMissed,
-        consecutiveMissed: randomConsecutiveMissed
     };
 }
